x86: separate out x86 build targets from base
Having extra x86_64/i686 targets in the base just makes the overall
size for other images bigger than it needs to be.  Instead, we just
put the script for installing rust in the image as part of base along
with basic system dependencies.

We add a new image, `:x86` for doing builds targetting that architecture
in its various forms.

Signed-off-by: Paul Osborne <osbpau@gmail.com>